+++ src/main/java/fr/raksrinana/fallingtree/tree/Tree.java
@@ -2,20 +2,18 @@ package fr.raksrinana.fallingtree.tree;
 
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IWorld;
+import net.minecraft.world.World;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.HashSet;
-import java.util.LinkedHashSet;
-import java.util.Objects;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Stream;
 
 public class Tree{
-	private final IWorld world;
+	private final World world;
 	private final Set<BlockPos> logs;
 	private final BlockPos hitPos;
 	
-	public Tree(@Nonnull IWorld world, @Nonnull BlockPos blockPos){
+	public Tree(@Nonnull World world, @Nonnull BlockPos blockPos){
 		this.world = world;
 		this.hitPos = blockPos;
 		this.logs = new LinkedHashSet<>();
@@ -37,12 +35,12 @@ public class Tree{
 	}
 	
 	@Nonnull
-	public IWorld getWorld(){
+	public World getWorld(){
 		return this.world;
 	}
 	
 	@Nonnull
-	public Stream<BlockPos> getLogs(){
-		return this.logs.stream();
+	public Collection<BlockPos> getLogs(){
+		return this.logs;
 	}
 }